refactor: 使用 .NET 5 引入的检测操作系统版本的方法

+++ Runtime/Utils.cpp
@@ -10,7 +10,7 @@ BOOL CALLBACK EnumChildProc(
 	_In_ LPARAM lParam
 ) {
 	std::wstring className(256, 0);
-	int num = GetClassName(hwnd, &className[0], className.size());
+	int num = GetClassName(hwnd, &className[0], (int)className.size());
 	if (num == 0) {
 		SPDLOG_LOGGER_ERROR(logger, MakeWin32ErrorMsg("GetClassName 失败"));
 		return TRUE;
